test: move api tests
