+++ gravitee-apim-e2e/api-test/src/apis/mapi-v1/api-plan.spec.ts
@@ -21,35 +21,35 @@ import {
   forPortalAsSimpleUser,
 } from '@gravitee/utils/configuration';
 import { afterAll, beforeAll, describe, expect } from '@jest/globals';
-import { APIsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIsApi';
-import { ApiEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApiEntity';
-import { ApiLifecycleState } from '@gravitee/management-webclient-sdk/src/lib/models/ApiLifecycleState';
+import { APIsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIsApi';
+import { ApiEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiEntity';
+import { ApiLifecycleState } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiLifecycleState';
 import { ApisFaker } from '@gravitee/fixtures/management/ApisFaker';
-import { Visibility } from '@gravitee/management-webclient-sdk/src/lib/models/Visibility';
+import { Visibility } from '../../../../lib/management-webclient-sdk/src/lib/models/Visibility';
 import { GroupsFaker } from '@gravitee/fixtures/management/GroupsFaker';
-import { SearchableUser } from '@gravitee/management-webclient-sdk/src/lib/models/SearchableUser';
-import { RoleScope } from '@gravitee/management-webclient-sdk/src/lib/models/RoleScope';
-import { GroupEntity } from '@gravitee/management-webclient-sdk/src/lib/models/GroupEntity';
-import { ConfigurationApi } from '@gravitee/management-webclient-sdk/src/lib/apis/ConfigurationApi';
-import { UsersApi } from '@gravitee/management-webclient-sdk/src/lib/apis/UsersApi';
+import { SearchableUser } from '../../../../lib/management-webclient-sdk/src/lib/models/SearchableUser';
+import { RoleScope } from '../../../../lib/management-webclient-sdk/src/lib/models/RoleScope';
+import { GroupEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/GroupEntity';
+import { ConfigurationApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/ConfigurationApi';
+import { UsersApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/UsersApi';
 import { find, cloneDeep } from 'lodash';
-import { ApplicationsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/ApplicationsApi';
+import { ApplicationsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/ApplicationsApi';
 import { ApplicationsFaker } from '@gravitee/fixtures/management/ApplicationsFaker';
-import { ApplicationEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApplicationEntity';
-import { ApiApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/ApiApi';
+import { ApplicationEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApplicationEntity';
+import { ApiApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/ApiApi';
 import { succeed } from '@lib/jest-utils';
 import { PlansFaker } from '@gravitee/fixtures/management/PlansFaker';
-import { PlanEntity } from '@gravitee/management-webclient-sdk/src/lib/models/PlanEntity';
-import { APIPlansApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIPlansApi';
-import { PlanStatus } from '@gravitee/management-webclient-sdk/src/lib/models/PlanStatus';
-import { SubscriptionApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/SubscriptionApi';
-import { Subscription } from '@gravitee/portal-webclient-sdk/src/lib/models/Subscription';
-import { APISubscriptionsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APISubscriptionsApi';
-import { PlanSecurityType } from '@gravitee/management-webclient-sdk/src/lib/models/PlanSecurityType';
-import { PlanValidationType } from '@gravitee/management-webclient-sdk/src/lib/models/PlanValidationType';
-import { UpdatePlanEntity } from '@gravitee/management-webclient-sdk/src/lib/models/UpdatePlanEntity';
-import { SubscriptionStatus } from '@gravitee/management-webclient-sdk/src/lib/models/SubscriptionStatus';
-import { UpdateApiEntityFromJSON } from '@gravitee/management-webclient-sdk/src/lib/models/UpdateApiEntity';
+import { PlanEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanEntity';
+import { APIPlansApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIPlansApi';
+import { PlanStatus } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanStatus';
+import { SubscriptionApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/SubscriptionApi';
+import { Subscription } from '../../../../lib/portal-webclient-sdk/src/lib/models/Subscription';
+import { APISubscriptionsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APISubscriptionsApi';
+import { PlanSecurityType } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanSecurityType';
+import { PlanValidationType } from '../../../../lib/management-webclient-sdk/src/lib/models/PlanValidationType';
+import { UpdatePlanEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/UpdatePlanEntity';
+import { SubscriptionStatus } from '../../../../lib/management-webclient-sdk/src/lib/models/SubscriptionStatus';
+import { UpdateApiEntityFromJSON } from '../../../../lib/management-webclient-sdk/src/lib/models/UpdateApiEntity';
 
 const orgId = 'DEFAULT';
 const envId = 'DEFAULT';
