+++ gravitee-apim-e2e/api-test/src/apis/mapi-v1/api-quality.spec.ts
@@ -15,21 +15,21 @@
  */
 import { forManagementAsAdminUser, forManagementAsApiUser, forManagementAsSimpleUser } from '@gravitee/utils/configuration';
 import { afterAll, beforeAll, describe, expect } from '@jest/globals';
-import { PortalApi } from '@gravitee/management-webclient-sdk/src/lib/apis/PortalApi';
-import { ConfigurationApi } from '@gravitee/management-webclient-sdk/src/lib/apis/ConfigurationApi';
-import { UsersApi } from '@gravitee/management-webclient-sdk/src/lib/apis/UsersApi';
-import { SearchableUser } from '@gravitee/management-webclient-sdk/src/lib/models/SearchableUser';
-import { QualityRuleEntity } from '@gravitee/management-webclient-sdk/src/lib/models/QualityRuleEntity';
-import { RoleScope } from '@gravitee/management-webclient-sdk/src/lib/models/RoleScope';
-import { GroupEntity } from '@gravitee/management-webclient-sdk/src/lib/models/GroupEntity';
-import { APIsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIsApi';
-import { ApiEntity, ApiEntityStateEnum } from '@gravitee/management-webclient-sdk/src/lib/models/ApiEntity';
-import { WorkflowState } from '@gravitee/management-webclient-sdk/src/lib/models/WorkflowState';
-import { ApiLifecycleState } from '@gravitee/management-webclient-sdk/src/lib/models/ApiLifecycleState';
-import { LifecycleAction } from '@gravitee/management-webclient-sdk/src/lib/models/LifecycleAction';
-import { ReviewAction } from '@gravitee/management-webclient-sdk/src/lib/models/ReviewAction';
+import { PortalApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/PortalApi';
+import { ConfigurationApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/ConfigurationApi';
+import { UsersApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/UsersApi';
+import { SearchableUser } from '../../../../lib/management-webclient-sdk/src/lib/models/SearchableUser';
+import { QualityRuleEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/QualityRuleEntity';
+import { RoleScope } from '../../../../lib/management-webclient-sdk/src/lib/models/RoleScope';
+import { GroupEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/GroupEntity';
+import { APIsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIsApi';
+import { ApiEntity, ApiEntityStateEnum } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiEntity';
+import { WorkflowState } from '../../../../lib/management-webclient-sdk/src/lib/models/WorkflowState';
+import { ApiLifecycleState } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiLifecycleState';
+import { LifecycleAction } from '../../../../lib/management-webclient-sdk/src/lib/models/LifecycleAction';
+import { ReviewAction } from '../../../../lib/management-webclient-sdk/src/lib/models/ReviewAction';
 import { fail, succeed } from '@lib/jest-utils';
-import { CurrentUserApi } from '@gravitee/management-webclient-sdk/src/lib/apis/CurrentUserApi';
+import { CurrentUserApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/CurrentUserApi';
 import { ApisFaker } from '@gravitee/fixtures/management/ApisFaker';
 import { GroupsFaker } from '@gravitee/fixtures/management/GroupsFaker';
 import { find } from 'lodash';
