+++ gravitee-apim-e2e/api-test/src/apis/mapi-v1/api-visibility.spec.ts
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { APIsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIsApi';
-import { ApiApi } from '@gravitee/portal-webclient-sdk/src/lib/apis/ApiApi';
+import { APIsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIsApi';
+import { ApiApi } from '../../../../lib/portal-webclient-sdk/src/lib/apis/ApiApi';
 import {
   forManagementAsApiUser,
   forPortalAsAdminUser,
@@ -25,15 +25,15 @@ import {
 } from '@gravitee/utils/configuration';
 import { afterAll, beforeAll, describe, expect } from '@jest/globals';
 import { ApisFaker } from '@gravitee/fixtures/management/ApisFaker';
-import { ApiEntity } from '@gravitee/management-webclient-sdk/src/lib/models/ApiEntity';
+import { ApiEntity } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiEntity';
 import { fail, succeed } from '@lib/jest-utils';
-import { ApiLifecycleState } from '@gravitee/management-webclient-sdk/src/lib/models/ApiLifecycleState';
-import { SearchableUser } from '@gravitee/management-webclient-sdk/src/lib/models/SearchableUser';
-import { UsersApi } from '@gravitee/management-webclient-sdk/src/lib/apis/UsersApi';
+import { ApiLifecycleState } from '../../../../lib/management-webclient-sdk/src/lib/models/ApiLifecycleState';
+import { SearchableUser } from '../../../../lib/management-webclient-sdk/src/lib/models/SearchableUser';
+import { UsersApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/UsersApi';
 import { find } from 'lodash';
-import { APIMembershipsApi } from '@gravitee/management-webclient-sdk/src/lib/apis/APIMembershipsApi';
-import { UpdateApiEntity, UpdateApiEntityFromJSON } from '@gravitee/management-webclient-sdk/src/lib/models/UpdateApiEntity';
-import { Visibility } from '@gravitee/management-webclient-sdk/src/lib/models/Visibility';
+import { APIMembershipsApi } from '../../../../lib/management-webclient-sdk/src/lib/apis/APIMembershipsApi';
+import { UpdateApiEntity, UpdateApiEntityFromJSON } from '../../../../lib/management-webclient-sdk/src/lib/models/UpdateApiEntity';
+import { Visibility } from '../../../../lib/management-webclient-sdk/src/lib/models/Visibility';
 
 const orgId = 'DEFAULT';
 const envId = 'DEFAULT';
